+++ CHANGES.md
@@ -1,6 +1,7 @@
 ### Next version
 * Convert to typescript project
 * Add share-conversion via `catalog-converter
+* Add basic multi-tenant config behaviour
 * [The next feature]
 
 ### 3.3.1